Menu and dropdown done.

+++ app/controllers/order_items_controller.rb
@@ -1,7 +1,17 @@
 class OrderItemsController < ApplicationController
   def destroy
     order_item = OrderItem.find(params[:id])
+    order = order_item.order
     authorize order_item
-    order_item.destroy
+    if order_item.quantity > 1
+      order_item.quantity -= 1
+      order.total_price -= order_item.item.price.to_d
+      order.save!
+      order_item.save!
+    else
+      order.total_price -= order_item.item.price.to_d
+      order.save!
+      order_item.destroy
+    end
   end
 end
